logic beres

+++ database/migrations/2024_01_25_012743_create_barangs_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->index('id_barang');
             $table->string('nama_produk')->unique();
             $table->string('merk')->nullable();
-            $table->integer('harga');
+            $table->bigInteger('harga');
             $table->string('image')->nullable();
             $table->timestamps();
         });
